domains ディレクトリを廃止した
characters.ts はsrc直下に移動した。path alias の設定変更も反映している。
