--- react/src/domains/character/index.ts
@@ -1,1 +1,0 @@
-export { characters } from './character';
