--- react/src/domains/index.ts
@@ -1,1 +1,0 @@
-export { characters } from './character';
